Rename property dir to Property
